Revert snake case file rename.
Changing the file name breaks compatibility and would cause issues for
projects already using this gem.
